// src/api/model/productCard.ts
/**
 * Generated by orval v7.4.1 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */
import type { ProductGroupRef } from './productGroupRef';
import type { UnitEnum } from './unitEnum';
import type { ProductTypeEnum } from './productTypeEnum';
import type { ProductCardUnit } from './productCardUnit';

export interface ProductCard {
  /** @nullable */
  id?: number | null;
  productCode: string;
  productName: string;
  productGroup: ProductGroupRef;
  defaultUnit: UnitEnum;
  productType: ProductTypeEnum;
  /** @nullable */
  optimalLevel?: number | null;
  /** @nullable */
  minimumLevel?: number | null;
  /** @nullable */
  specialCode?: string | null;
  /** @nullable */
  shelfLocation?: string | null;
  /** @nullable */
  productCardUnits?: ProductCardUnit[] | null;
}